requirements.txt
pyTelegramBotAPI==4.13.0
apscheduler==3.10.0
python-dotenv==0.21.1
pyyaml
aiohttp>=3.8.0
tzdata
